core/admin/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-bootstrap
Flask-Babel
Flask-migrate
Flask-script
Flask-wtf
Flask-debugtoolbar
Flask-limiter
redis
WTForms-Components
passlib
gunicorn
tabulate
PyYAML
PyOpenSSL
dnspython
bcrypt
tenacity
mysqlclient
psycopg2
